backend/requirements.txt
fastapi==0.133.1
uvicorn[standard]==0.41.0
gunicorn==25.1.0
vectorbt==0.28.4
duckdb==1.4.4
pandas==2.3.3
numpy==2.2.6
pandas_ta==0.4.71b0
python-dotenv==1.2.1
pydantic==2.12.5
